feat : 청원 답변 조회 기능 추가

+++ data/src/main/java/com/marassu/data/source/PetitionAnswerRemoteDataSource.kt
@@ -15,7 +15,7 @@ class PetitionAnswerRemoteDataSource(
     suspend fun postPetitionAnswer(petitionAnswerRequest: PetitionAnswerRequest) =
         tokenPetitionAnswerService.postPetitionAnswer(petitionAnswerRequest)
 
-    suspend fun getPetitionAnswer(petitionId: Int) =
+    suspend fun getPetitionAnswer(petitionId: Long) =
         publicPetitionAnswerService.getPetitionAnswer(petitionId)
 
     suspend fun deletePetitionAnswer(petitionId: Int) =
